added merge_tables and link_tables function to trp2.py. The test functions also added to test_trp2.py.
merge_table functions takes an array of table_ids. It assumes the first table as the parent table and extends its cells with the cells of the other tables.
link_table functions takes an array of table_ids and create custom fields (next_table and previous_table) pointing to the the next or previous tables ids in the input array.