fix(js): avoid Array.flat() in getLinesByLayoutArea
Seems like .flat() is missing from the Array class in some versions
of Node that we'd like to support, and not getting caught by TS
compilation. Switch to destructuring-style flattening (which we were
already using in other places) for now, and hopefully we can fix it
in the compilation toolchain sometime in future.