fix(js): Use CJS instead of IIFE for webpack
Remove the `browser` field from package.json which previously pointed
to the IIFE build of the library, and replace it with a `jsdelivr`
field pointing to same location. While IIFE is appropriate for
direct-to-browser use like with CDNs, front end frameworks like
webpack also refer to this field and should use standard CommonJS
build instead.