fix(js): Support 'KEY' & 'VALUE' blocks for forms
In some specific test data files, we observed Textract reporting
distinct 'KEY' and 'VALUE' block types for forms KV data instead of
the usual 'KEY_VALUE_SET'. Not sure whether this was a transient API
issue or something we can expect to see more of, but added support
for the alternative option in case.