fix(js): Workaround for TypeScript mixin errors
Although 0.4.0-alpha.1 tested and built fine, TypeScript project
users were finding errors when attempting to use TRP.js because of
invalid generated type declarations due to open TypeScript issue
54879. This temporary workaround uses a CLI find and replace tool to
directly edit the generated declarations, and is predicated on the
assumption that we never use non-getter `.text` properties (the
replace is case-sensitive so doesn't corrupt the API models' `.Text`
properties). It's a hacky fix, and would be keen to remove as soon as
the upstream TypeScript issue is resolved.